// packages/ide/src/types/trace-types.ts
/**
 *
 * Copyright (c) 2026 Analog Devices, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */

export type TraceConfiguration = {
  interfaceType: "uart" | "usb";
  serialPort: string;
  baudRate: number;
  outputDirectory: string;
  autoReset: boolean;

  elfFile: string;
  buildDir: string;
  // Ai model file paths
  aiModels: string[];
};


export type TraceRecordingState =
  | {
      isRecording: true;
      startTime: number;
    }
  | {
      isRecording: false;
      error?: string;
    };

export type SerialPort = {
  path: string;
};